refactor: feed, feed comment securityUtil 적용, QueryDSL 수정

+++ src/main/java/com/team1/spreet/domain/feed/service/FeedCommentService.java
@@ -8,6 +8,7 @@ import com.team1.spreet.domain.feed.repository.FeedRepository;
 import com.team1.spreet.domain.user.model.User;
 import com.team1.spreet.global.error.exception.RestApiException;
 import com.team1.spreet.global.error.model.ErrorStatusCode;
+import com.team1.spreet.global.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,12 +23,20 @@ public class FeedCommentService {
     private final FeedCommentRepository feedCommentRepository;
 
     @Transactional
-    public void saveFeedComment(Long feedId, FeedCommentDto.RequestDto requestDto, User user) {
+    public void saveFeedComment(Long feedId, FeedCommentDto.RequestDto requestDto) {
+        User user = SecurityUtil.getCurrentUser();
+        if(user == null){
+            throw new RestApiException(ErrorStatusCode.NOT_EXIST_AUTHORIZATION);
+        }
         Feed feed = checkFeed(feedId);
         feedCommentRepository.saveAndFlush(requestDto.toEntity(feed, user));
     }
     @Transactional
-    public void updateFeedComment(Long commentId, FeedCommentDto.RequestDto requestDto, User user){
+    public void updateFeedComment(Long commentId, FeedCommentDto.RequestDto requestDto){
+        User user = SecurityUtil.getCurrentUser();
+        if(user == null){
+            throw new RestApiException(ErrorStatusCode.NOT_EXIST_AUTHORIZATION);
+        }
         FeedComment feedComment = checkFeedComment(commentId);
         if(checkOwner(feedComment, user.getId())){
             feedComment.update(requestDto.getContent());
@@ -35,7 +44,11 @@ public class FeedCommentService {
     }
 
     @Transactional
-    public void deleteFeedComment(Long commentId, User user) {
+    public void deleteFeedComment(Long commentId) {
+        User user = SecurityUtil.getCurrentUser();
+        if(user == null){
+            throw new RestApiException(ErrorStatusCode.NOT_EXIST_AUTHORIZATION);
+        }
         FeedComment feedComment = checkFeedComment(commentId);    //userId, commentId로 comment 찾기
         if(checkOwner(feedComment, user.getId())) {
             feedComment.delete();
